+++ src/layers/dense.hpp
@@ -2,6 +2,7 @@
 
 #include "activations.hpp"
 #include "../utils.hpp"
+#include <layers.pb.h>
 
 
 class DenseLayer : public ILayer {
@@ -60,6 +61,29 @@ public:
                                }
                 }
         };
+    }
+
+    void SaveWeights(std::ostream *out) {
+        DenseWeights matrix;
+        matrix.set_n_rows(weights_and_bias.n_rows);
+        matrix.set_n_cols(weights_and_bias.n_cols);
+        for (arma::uword i = 0; i < weights_and_bias.n_rows; ++i) {
+            auto row = matrix.add_vectors();
+            for (arma::uword j = 0; j < weights_and_bias.n_cols; ++j) {
+                row->add_scalars(weights_and_bias.at(i, j));
+            }
+        }
+        matrix.SerializeToOstream(out);
+    }
+
+    void LoadWeights(std::istream *in) {
+        DenseWeights matrix;
+        matrix.ParseFromIstream(in);
+        for (::google::protobuf::uint32 i = 0; i < matrix.n_rows(); ++i) {
+            for (::google::protobuf::uint32 j = 0; j < matrix.n_cols(); ++j) {
+                weights_and_bias.at(i, j) = matrix.vectors(i).scalars(j);
+            }
+        }
     }
 
 private:
